fix(front): repos and persons not updating
