+++ src/routes/dashboard/+page.svelte
@@ -5,21 +5,8 @@
 	import { page } from '$app/stores';
 	import { Error, Main, Sidebar } from '~/lib/components';
 	import { createNotificationData, fetchGithub, storage } from '~/lib/helpers';
-	import {
-		error,
-		githubNotifications,
-		loading,
-		savedNotifications,
-		settings,
-		watchedPersons,
-		watchedRepos
-	} from '~/lib/stores';
-	import type {
-		GithubNotification,
-		NotificationData,
-		WatchedPerson,
-		WatchedRepo
-	} from '~/lib/types';
+	import { error, githubNotifications, loading, savedNotifications, settings } from '~/lib/stores';
+	import type { GithubNotification, NotificationData } from '~/lib/types';
 
 	const user = $page.data.session?.user;
 
@@ -101,53 +88,6 @@
 			...newNotifications,
 			...$githubNotifications.filter((item) => !newNotifications.find((n) => n.id === item.id))
 		];
-
-		// Update watched repos
-		const savedWatchedRepos = storage.get('github-watched-repos');
-		$watchedRepos = $githubNotifications.reduce<WatchedRepo[]>((previous, current) => {
-			if (current.done) return previous;
-			const index = previous.findIndex((repo) => repo.id === current.repoId);
-			if (index > -1) {
-				const repo = previous.splice(index, 1)[0];
-				return [...previous, { ...repo, number: repo.number + 1 }];
-			}
-			return [
-				...previous,
-				{
-					id: current.repoId,
-					name: current.repo,
-					ownerName: current.owner,
-					ownerAvatar: current.ownerAvatar,
-					number: 1,
-					active: savedWatchedRepos?.find((repo) => repo.id === current.repoId)?.active ?? true
-				}
-			];
-		}, []);
-
-		// Update watched persons
-		const savedWatchedPersons = storage.get('github-watched-persons');
-		$watchedPersons = $githubNotifications
-			.reduce<WatchedPerson[]>((previous, current) => {
-				if (!current.author || current.done) return previous;
-				const index = previous.findIndex((person) => person.login === current?.author?.login);
-				if (index > -1) {
-					const person = previous.splice(index, 1)[0];
-					return [...previous, { ...person, number: person.number + 1 }];
-				}
-				return [
-					...previous,
-					{
-						login: current.author?.login ?? '',
-						avatar: current.author?.avatar ?? '',
-						number: 1,
-						bot: current.author.bot,
-						active:
-							savedWatchedPersons?.find((person) => person.login === current.author?.login)
-								?.active ?? true
-					}
-				];
-			}, [])
-			.sort((a, b) => b.number - a.number);
 	}
 
 	function refetch() {
